Configuration Wizard: Column editor add action "Add and New"
This patch improves the UX when adding new columns in an entity
by allowing you to add several new columns without closing the
column editor window.

+++ stdm/ui/wizard/column_editor.py
@@ -199,7 +199,7 @@ class ColumnEditor(WIDGET, BASE):
 
     def init_controls(self):
         """
-        Initialize GUI controls default state when the dialog window is opened.
+        Initialize GUI controls 
         """
         self.populate_data_type_cbo()
 
@@ -208,24 +208,20 @@ class ColumnEditor(WIDGET, BASE):
             self.column_to_wa(self.column)
 
         self.edtColName.setFocus()
-
         self.edtColName.setEnabled(not self.in_db)
-
         self.cboDataType.setEnabled(not self.in_db)
 
-        self.buttonBox.button(QDialogButtonBox.Ok).clicked.connect(self.accept)
-        self.buttonBox.button(QDialogButtonBox.Cancel).clicked.connect(self.cancel)
+        self.btnAdd.clicked.connect(self.accept)
+        self.btnAddNew.clicked.connect(self.add_new_column)
+        self.btnClose.clicked.connect(self.cancel)
 
         col_type = self._column_type_info(self.column)
+
         if not self.in_db and col_type == 'GEOMETRY':
             opts = self.type_attribs[col_type]
             self.cbMandt.setEnabled(opts['mandt']['enabled_state'])
             self.cbUnique.setEnabled(opts['unique']['enabled_state'])
             self.cbIndex.setEnabled(opts['index']['enabled_state'])
-        # else:
-        #     self.cbMandt.setEnabled(not self.in_db)
-        #     self.cbUnique.setEnabled(not self.in_db)
-        #     self.cbIndex.setEnabled(not self.in_db)
 
         # Dont allow mandatory fields if an entity already has records.
         # if self.entity_has_records:
@@ -927,7 +923,25 @@ class ColumnEditor(WIDGET, BASE):
         msg.setText(message)
         msg.exec_()
 
-    def accept(self):
+    def add_new_column(self):
+        new_column = self.make_column()
+
+        if not self.column_is_valid(new_column):
+            return
+
+        self.form_parent.add_new_column(new_column)
+        self.clear_form_fields()
+
+    def clear_form_fields(self):
+        self.edtColName.setText('')
+        self.edtColDesc.setText('')
+        self.txt_form_label.setText('')
+        self.edtUserTip.setText('')
+        self.cbMandt.setCheckState(Qt.Unchecked)
+        self.cbSearch.setCheckState(Qt.Unchecked)
+        self.cbUnique.setCheckState(Qt.Unchecked)
+
+    def column_is_valid(self, column)->bool:
         col_name = str(self.edtColName.text()).strip()
         # column name is not empty
         if len(col_name) == 0 or col_name == '_':
@@ -941,37 +955,38 @@ class ColumnEditor(WIDGET, BASE):
                         "Please choose another column name.".format(col_name)))
             return False
 
-        new_column = self.make_column()
-
-        if new_column is None:
-            LOGGER.debug("Error creating column!")
-            self.show_message('Unable to create column!')
-            return False
-
-        if self.cbMandt.isChecked():
-            if self.entity_has_null_values():
-                self.show_message(self.tr("Cannot set column as mandatory. "
-                                            "Entity has null values!"))
-                return False
-
-        if self.cbUnique.isChecked():
-            if self.column_has_no_unique_values():
-                self.show_message(self.tr("Cannot set column as unique. "
-                                          "Entity has non-unique values!"))
-                return False
-
         if self.column is None:  # new column
             if self.column_exists(col_name):
                 self.show_message(self.tr("Column with the same name already "
                                           "exist in this entity!"))
                 return False
 
-            if self.auto_entity_add:
-                self.entity.add_column(new_column)
+        return True
 
-            self.column = new_column
+    def accept(self):
+        new_column = self.make_column()
+        self.column = new_column
 
-        else:  # editing a column
+        if not self.column_is_valid(new_column):
+            return
+
+        if self.auto_entity_add:
+            self.entity.add_column(new_column)
+
+        if not self.is_new:  # editing a column
+
+            if self.cbMandt.isChecked():
+                if self.entity_has_null_values(self.column):
+                    self.show_message(self.tr("Cannot set column as mandatory. "
+                                                "Entity has null values!"))
+                    return False
+
+            if self.cbUnique.isChecked():
+                if self.column_has_no_unique_values():
+                    self.show_message(self.tr("Cannot set column as unique. "
+                                            "Entity has non-unique values!"))
+                    return False
+
             if isinstance(self.prev_column, ForeignKeyColumn):
                 if self.prev_column.display_name() != new_column.display_name():
                     self.entity.remove_column(self.prev_column.name)
@@ -980,12 +995,13 @@ class ColumnEditor(WIDGET, BASE):
 
         self.done(1)
 
-    def entity_has_null_values(self)->bool:
-        sql_stmt = f'Select count(*) cnt from {self.column.entity.name} where {self.column.name} is null'  
+    def entity_has_null_values(self, column: 'Column')->bool:
+        sql_stmt = f'Select count(*) cnt from {column.entity.name} where {column.name} is null'  
         results = run_query(sql_stmt)
         cnt = 0
         for result in results:
             cnt = result['cnt']
+
         return True if cnt> 0 else False
 
     def column_has_no_unique_values(self)->bool:
